requirements.txt
graphviz>=0.17
pytest>=6.2.5
pyyaml>=6.0
pyreadline3>=3.4.1
sympy>=1.11.1
termcolor>=1.1.0